asl-to-text-ai/requirements.txt
# Core ML and Computer Vision (Render-optimized)
tensorflow-cpu==2.15.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
scipy==1.11.4

# Web Framework and API
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
requests==2.31.0

# Video Processing
imageio==2.31.5
pillow==10.0.1

# Data Processing
pandas==2.1.3
joblib==1.3.2

# Deployment
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
